+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [6.0.10] - 2022-02-13
+### Added
+- An input field for the maximum number of recent Observations per Patient
+  to check when no code is specified in the criteria.
+
 ## [6.0.9] - 2022-02-06
 ### Changed
 - Temporarily disabled the use of the _security URL parameter.